+++ include/picongpu/fields/absorber/pml/Solver.kernel
@@ -22,7 +22,7 @@
 
 #include "picongpu/simulation_defines.hpp"
 
-#include "picongpu/fields/MaxwellSolver/YeePML/Parameters.hpp"
+#include "picongpu/fields/absorber/pml/Parameters.hpp"
 
 #include <pmacc/algorithms/math/floatMath/floatingPoint.tpp>
 #include <pmacc/mappings/threads/ForEachIdx.hpp>
@@ -38,9 +38,9 @@ namespace picongpu
 {
     namespace fields
     {
-        namespace maxwellSolver
+        namespace absorber
         {
-            namespace yeePML
+            namespace pml
             {
                 //! Parameters of PML for the local domain
                 struct LocalParameters : public Parameters
@@ -463,7 +463,7 @@ namespace picongpu
                     }
                 };
 
-            } // namespace yeePML
-        } // namespace maxwellSolver
+            } // namespace pml
+        } // namespace absorber
     } // namespace fields
 } // namespace picongpu
